ai-service/requirements.txt
# FILE: ai-service/requirements.txt
python-multipart
torch
diffusers[torch]>=0.20.0
transformers
accelerate
Pillow
runpod